+++ HISTORY.rst
@@ -15,6 +15,15 @@ Recent History
 .. Next Release
 .. ------------------
 
+1.4.1 (2021-02-11)
+------------------
+
+* Add accounting code from title to list of event tags.
+* Rename "Meetup RSVP" registration type to "Instructor/Host".
+* Disable guest registration due to COVID-19.
+* Report and limit oversubscribed Meetup events.
+* Add -m flag to show Meetup event IDs in reports.
+
 1.4.0 (2020-11-29)
 ------------------